requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
Pillow==10.2.0
pytest==7.4.3
pytest-django==4.5.2
django-extensions==3.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
